# Conflicts:
#	lib/bloc/company_products/company_products_bloc.dart
#	lib/bloc/home/home_bloc.dart
#	lib/bloc/menu/menu_bloc.dart
#	lib/bloc/planogram_product/planogram_product_bloc.dart
#	lib/bloc/product_details/product_details_bloc.dart
#	lib/bloc/store/store_bloc.dart
#	lib/bloc/store_category/store_category_bloc.dart
#	lib/bloc/supplier_products/supplier_products_bloc.dart
#	lib/ui/screens/company_product_screen.dart
#	lib/ui/screens/planogram_product_screen.dart
#	lib/ui/screens/recommendation_products_screen.dart
#	lib/ui/screens/reorder_screen.dart
#	lib/ui/screens/supplier_products_screen.dart
#	lib/ui/utils/themes/app_urls.dart

+++ lib/data/model/res_model/previous_order_products_res_model/previous_order_products_res_model.dart
@@ -34,7 +34,7 @@ class PreviousOrderProductsResModel with _$PreviousOrderProductsResModel {
 class PreviousOrderProductData with _$PreviousOrderProductData {
   const factory PreviousOrderProductData({
     @JsonKey(name: "_id") String? id,
-    @JsonKey(name: "productStock") int? productStock,
+    @JsonKey(name: "productStock") double? productStock,
     @JsonKey(name: "totalSale") int? totalSale,
     @JsonKey(name: "productPrice") double? productPrice,
     @JsonKey(name: "mainImage") String? mainImage,
